+++ changelog.md
@@ -1,3 +1,7 @@
+### 1.0.2
+- Added presets (The old `on_block_hit` structure is still valid)
+- Field `tag` in `on_block_hit` now optional
+
 ### 1.0.1
 - Added `break_steps` field
 - Added Icon